backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
motor==3.4.0
pymongo==4.7.2
pydantic[email]==2.7.1
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
anthropic>=0.26.0
python-dotenv==1.0.1
pytest==8.2.0
pytest-asyncio==0.23.7
httpx==0.27.0
